+++ src/Validation/src/Rules/Errors/ErrorTemplateRegistry.php
@@ -33,15 +33,7 @@ class ErrorTemplateRegistry
      */
     public function getErrorTemplate(string $field, string $ruleSlug): string
     {
-        if (isset($this->fieldTemplates[$field][$ruleSlug])) {
-            return $this->fieldTemplates[$field][$ruleSlug];
-        }
-
-        if (isset($this->globalTemplates[$ruleSlug])) {
-            return $this->globalTemplates[$ruleSlug];
-        }
-
-        return '';
+        return $this->fieldTemplates[$field][$ruleSlug] ?? $this->globalTemplates[$ruleSlug] ?? '';
     }
 
     /**